refactoring: do more flat structure into screens folder
