+++ src/App.js
@@ -1,14 +1,14 @@
 import React, { useEffect } from 'react';
-import LoginPage from './screens/Login/LoginPage/LoginPage';
+import Login from './screens/Login/Login';
 import { Routes, Route, } from 'react-router-dom';
 import RequireAuth from './hoc/RequireAuth';
 import Layout from './layouts/Layout/Layout';
-import HomePage from './screens/Home/HomePage/HomePage';
-import PeoplePage from './screens/People/People-page/People-page';
-import SettingsPage from './screens/SettingsPage/SettingsPage';
-import PersonDetailsPage from './screens/PersonDetails/PersonDetailsPage/PersonDetailsPage';
-import ColorsPage from './screens/Colors/ColorsPage/ColorsPage';
-import ColorDetailsPage from './screens/ColorDetails/ColorDetailsPage/ColorDetailsPage';
+import Home from './screens/Home/Home';
+import People from './screens/People/People';
+import Settings from './screens/Settings/Settings';
+import PersonDetails from './screens/PersonDetails/PersonDetails';
+import Colors from './screens/Colors/Colors';
+import ColorDetails from './screens/ColorDetails/ColorDetails';
 import { useDispatch } from 'react-redux';
 import loaderController from './interceptors/loaderController';
 import { ROUTES } from './enums/ROUTES';
@@ -29,16 +29,16 @@ function App() {
             <Layout />
           </RequireAuth>
         }>
-          <Route index element={<HomePage />} />
-          <Route path={`${ROUTES.colors}/:page`} element={<ColorsPage />} />
-          <Route path={`${ROUTES.color}/:id`} element={<ColorDetailsPage />} />
-          <Route path={`${ROUTES.people}/:page`} element={<PeoplePage />} />
-          <Route path={`${ROUTES.person}/:id`} element={<PersonDetailsPage />}/>
-          <Route path={ROUTES.settings} element={<SettingsPage />} />
-          <Route path={ROUTES.notFound} element={<HomePage />} />
+          <Route index element={<Home />} />
+          <Route path={`${ROUTES.colors}/:page`} element={<Colors />} />
+          <Route path={`${ROUTES.color}/:id`} element={<ColorDetails />} />
+          <Route path={`${ROUTES.people}/:page`} element={<People />} />
+          <Route path={`${ROUTES.person}/:id`} element={<PersonDetails />}/>
+          <Route path={ROUTES.settings} element={<Settings />} />
+          <Route path={ROUTES.notFound} element={<Home />} />
         </Route>
 
-        <Route path={ROUTES.authorization} element={<LoginPage />} />
+        <Route path={ROUTES.authorization} element={<Login />} />
       </Routes>
     </>
   );
