+++ src/screens/ColorDetails/ColorDetails.module.scss
