+++ src/screens/Login/LoginForm/LoginForm.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import styles from './Login-form.module.scss';
-import AuthForm from '../Auth-form/Auth-form';
+import styles from './LoginForm.module.scss';
+import AuthForm from '../AuthForm/AuthForm';
 import StringButton from '../../../components/shared/String-button/String-button';
 import Popup from '../../../components/Popup/Popup';
