+++ src/screens/People/People.module.scss
