+++ src/screens/People/PersonCard/PersonCard.module.scss
