+++ src/screens/PersonDetails/PersonDetails.module.scss
