+++ src/screens/Settings/Settings.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+// import styles from './Settings.module.scss';
+
+const Settings = () => {
+  return (
+    <h1>Settings Page works!!!</h1>
+  )
+}
+
+export default Settings;
