--- src/screens/SettingsPage/SettingsPage.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-// import styles from './SettingsPage.module.scss';
-
-const SettingsPage = () => {
-  return (
-    <h1>Settings Page works!!!</h1>
-  )
-}
-
-export default SettingsPage;
